split TypeAlias and TypeAnnotation to separate visitors
